+++ src/app/modules.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
 import { Application } from 'lisk-sdk';
-import { BuyerModule } from './modules/buyer/buyer_module';
-import { SellerModule } from './modules/seller/seller_module';
+import { BuyerModule } from './modules/buyer/buyerModule';
+import { SellerModule } from './modules/seller/sellerModule';
 
 export const registerModules = (app: Application): void => {
   app.registerModule(SellerModule);
