+++ src/app/plugins.ts
@@ -2,10 +2,10 @@
 import { DashboardPlugin } from '@liskhq/lisk-framework-dashboard-plugin';
 import { FaucetPlugin } from '@liskhq/lisk-framework-faucet-plugin';
 import { Application } from 'lisk-sdk';
-import { BazarrestapiPlugin } from './plugins/bazarrestapi/bazarrestapi_plugin';
+import { BazarRestApiPlugin } from './plugins/bazarRestApi/bazarRestApiPlugin';
 
 export const registerPlugins = (app: Application): void => {
   app.registerPlugin(DashboardPlugin);
   app.registerPlugin(FaucetPlugin);
-  app.registerPlugin(BazarrestapiPlugin);
+  app.registerPlugin(BazarRestApiPlugin);
 };
