+++ src/app/plugins/bazarRestApi/controller/block/getBlockById.ts
