+++ src/app/plugins/bazarRestApi/controller/block/getLastBlock.ts
