+++ src/app/plugins/bazarRestApi/controller/peer/getConnectedPeers.ts
