+++ src/commands/hashOnion.ts
