+++ test/unit/plugins/bazarRestApi/bazarRestApiPlugin.spec.ts
